Minor improvements to the registration code

+++ src/main/resources/db/changelog/migrations/initial-db-setup.sql
@@ -1,6 +1,6 @@
 create table account
 (
-  id       numeric primary key,
+  id       serial primary key,
   username varchar(26)  not null unique,
   password varchar(255) not null,
   email    varchar(255) not null
